Int64 -> Int32; first iteration of improving the operations

+++ src/DataTypes/DataTypeLowCardinality.cpp
@@ -84,7 +84,7 @@ MutableColumnUniquePtr DataTypeLowCardinality::createColumnUniqueImpl(const IDat
     if (which.isDateTime())
         return creator(static_cast<ColumnVector<UInt32> *>(nullptr));
     if (which.isTime())
-        return creator(static_cast<ColumnVector<Int64> *>(nullptr));
+        return creator(static_cast<ColumnVector<Int32> *>(nullptr));
     if (which.isUUID())
         return creator(static_cast<ColumnVector<UUID> *>(nullptr));
     if (which.isIPv4())
